server: rename `network` dir to `web3`
