+++ packages/server/src/validators/RevealsValidator.js
@@ -1,5 +1,5 @@
 import Models from '../models'
-import Network from '../network/Network'
+import Network from '../web3/Network'
 import BaseValidator from './BaseValidator'
 const { hashVote } = require('@aragon/court/test/helpers/utils/crvoting')
